flask_api/movie-recommendation-api/requirements.txt
flask
flask-restful
mysql-connector-python
psycopg2-binary
passlib
email_validator
Flask-JWT-Extended